requirements.txt
openai>=1.0.0
anthropic>=0.21.0
google-generativeai>=0.3.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Token counting and evaluation dependencies
tiktoken>=0.5.0
ragas>=0.1.0
datasets>=2.14.0
